:hammer: Break up the gemtext widgets into smaller units
Has no impact on the workings of the application, just makes the code a
little easier to navigate.
